requirements.txt
langchain_google_genai
langchain_openai
langchain
langsmith
langchain_community
faker
python-dotenv
pandas
openpyxl